pyproject.toml
[tool.poetry]
name = "roombapy"
version = "1.6.12"
description = "Python program and library to control Wi-Fi enabled iRobot Roombas"
authors = ["Philipp Schmitt <philipp@schmitt.co>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/pschmitt/roombapy"
classifiers = [
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Networking"
]
packages = [
    { include = "roombapy" },
]

[tool.poetry.scripts]
roomba-connect = "roombapy.entry_points:connect"
roomba-discovery = "roombapy.entry_points:discovery"
roomba-password = "roombapy.entry_points:password"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
orjson = ">=3.9.13"
paho-mqtt = ">=1.5.1,<3.0.0"
pydantic = ">=1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
ruff = "^0.2.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
# Specify the target platform details in config, so your developers are
# free to run mypy on Windows, Linux, or macOS and get consistent
# results.
platform = "linux"
python_version = "3.10"

# show error messages from unrelated files
follow_imports = "normal"

# suppress errors about unsatisfied imports
ignore_missing_imports = true

# be strict
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
warn_incomplete_stub = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.ruff]
line-length = 80

[tool.ruff.lint]
ignore = [
    "ANN001", # Parameter type
    "ANN101", # Self... explanatory
    "ANN102", # cls... just as useless
    "ANN202",
    "ANN204", # init return type
    "ANN205",
    "ANN201", # Return type
    "ANN401", # Opinioated warning on disallowing dynamically typed expressions
    "ARG002",
    "BLE001",
    "B007",
    "C901",
    "D105",
    "D106",
    "D203", # Conflicts with other rules
    "D404",
    "D213", # Conflicts with other rules
    "DTZ005",
    "EM101",
    "EM102",
    "FBT001",
    "FBT002",
    "FBT003",
    "G002",
    "G004",
    "TRY002",
    "TRY003",
    "TRY300",
    "TRY400",
    "COM812", # Conflicts with other rules
    "ISC001", # Conflicts with other rules
    "D100", # documentation info, should be removed after fixes
    "D101", # documentation info, should be removed after fixes
    "D102", # documentation info, should be removed after fixes
    "D103", # documentation info, should be removed after fixes
    "PERF102",
    "PGH003",
    "PLE1205",
    "PLR2004", # Just annoying, not really useful
    "PLR0912",
    "PLR1714",
    "PLW2901",
    "RET507",
    "SIM102",
    "SLF001",
    "S101",
    "T201",
    "UP007",
]
select = ["ALL"]

[tool.ruff.format]
docstring-code-format = true